code working on endpoint query
